Move some ColumnFamily options to advanced_options.h
