+++ util/options_helper.h
@@ -365,6 +365,21 @@ static std::unordered_map<std::string, OptionTypeInfo> db_options_type_info = {
       OptionType::kBoolean, OptionVerificationType::kNormal, true,
       offsetof(struct MutableDBOptions, avoid_flush_during_shutdown)}}};
 
+// offset_of is used to get the offset of a class data member
+// ex: offset_of(&ColumnFamilyOptions::num_levels)
+// This call will return the offset of num_levels in ColumnFamilyOptions class
+//
+// This is the same as offsetof() but allow us to work with non standard-layout
+// classes and structures
+// refs:
+// http://en.cppreference.com/w/cpp/concept/StandardLayoutType
+// https://gist.github.com/graphitemaster/494f21190bb2c63c5516
+template <typename T1, typename T2>
+inline int offset_of(T1 T2::*member) {
+  static T2 obj;
+  return int(size_t(&(obj.*member)) - size_t(&obj));
+}
+
 static std::unordered_map<std::string, OptionTypeInfo> cf_options_type_info = {
     /* not yet supported
     CompactionOptionsFIFO compaction_options_fifo;
@@ -379,45 +394,44 @@ static std::unordered_map<std::string, OptionTypeInfo> cf_options_type_info = {
                                      std::string* merged_value);
      */
     {"report_bg_io_stats",
-     {offsetof(struct ColumnFamilyOptions, report_bg_io_stats),
-      OptionType::kBoolean, OptionVerificationType::kNormal, true,
+     {offset_of(&ColumnFamilyOptions::report_bg_io_stats), OptionType::kBoolean,
+      OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, report_bg_io_stats)}},
     {"compaction_measure_io_stats",
      {0, OptionType::kBoolean, OptionVerificationType::kDeprecated, false, 0}},
     {"disable_auto_compactions",
-     {offsetof(struct ColumnFamilyOptions, disable_auto_compactions),
+     {offset_of(&ColumnFamilyOptions::disable_auto_compactions),
       OptionType::kBoolean, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, disable_auto_compactions)}},
     {"filter_deletes",
      {0, OptionType::kBoolean, OptionVerificationType::kDeprecated, true, 0}},
     {"inplace_update_support",
-     {offsetof(struct ColumnFamilyOptions, inplace_update_support),
+     {offset_of(&ColumnFamilyOptions::inplace_update_support),
       OptionType::kBoolean, OptionVerificationType::kNormal, false, 0}},
     {"level_compaction_dynamic_level_bytes",
-     {offsetof(struct ColumnFamilyOptions,
-               level_compaction_dynamic_level_bytes),
+     {offset_of(&ColumnFamilyOptions::level_compaction_dynamic_level_bytes),
       OptionType::kBoolean, OptionVerificationType::kNormal, false, 0}},
     {"optimize_filters_for_hits",
-     {offsetof(struct ColumnFamilyOptions, optimize_filters_for_hits),
+     {offset_of(&ColumnFamilyOptions::optimize_filters_for_hits),
       OptionType::kBoolean, OptionVerificationType::kNormal, false, 0}},
     {"paranoid_file_checks",
-     {offsetof(struct ColumnFamilyOptions, paranoid_file_checks),
+     {offset_of(&ColumnFamilyOptions::paranoid_file_checks),
       OptionType::kBoolean, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, paranoid_file_checks)}},
     {"force_consistency_checks",
-     {offsetof(struct ColumnFamilyOptions, force_consistency_checks),
+     {offset_of(&ColumnFamilyOptions::force_consistency_checks),
       OptionType::kBoolean, OptionVerificationType::kNormal, false, 0}},
     {"purge_redundant_kvs_while_flush",
-     {offsetof(struct ColumnFamilyOptions, purge_redundant_kvs_while_flush),
+     {offset_of(&ColumnFamilyOptions::purge_redundant_kvs_while_flush),
       OptionType::kBoolean, OptionVerificationType::kNormal, false, 0}},
     {"verify_checksums_in_compaction",
      {0, OptionType::kBoolean, OptionVerificationType::kDeprecated, true, 0}},
     {"soft_pending_compaction_bytes_limit",
-     {offsetof(struct ColumnFamilyOptions, soft_pending_compaction_bytes_limit),
+     {offset_of(&ColumnFamilyOptions::soft_pending_compaction_bytes_limit),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, soft_pending_compaction_bytes_limit)}},
     {"hard_pending_compaction_bytes_limit",
-     {offsetof(struct ColumnFamilyOptions, hard_pending_compaction_bytes_limit),
+     {offset_of(&ColumnFamilyOptions::hard_pending_compaction_bytes_limit),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, hard_pending_compaction_bytes_limit)}},
     {"hard_rate_limit",
@@ -425,21 +439,21 @@ static std::unordered_map<std::string, OptionTypeInfo> cf_options_type_info = {
     {"soft_rate_limit",
      {0, OptionType::kDouble, OptionVerificationType::kDeprecated, true, 0}},
     {"max_compaction_bytes",
-     {offsetof(struct ColumnFamilyOptions, max_compaction_bytes),
+     {offset_of(&ColumnFamilyOptions::max_compaction_bytes),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_compaction_bytes)}},
     {"expanded_compaction_factor",
      {0, OptionType::kInt, OptionVerificationType::kDeprecated, true, 0}},
     {"level0_file_num_compaction_trigger",
-     {offsetof(struct ColumnFamilyOptions, level0_file_num_compaction_trigger),
+     {offset_of(&ColumnFamilyOptions::level0_file_num_compaction_trigger),
       OptionType::kInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, level0_file_num_compaction_trigger)}},
     {"level0_slowdown_writes_trigger",
-     {offsetof(struct ColumnFamilyOptions, level0_slowdown_writes_trigger),
+     {offset_of(&ColumnFamilyOptions::level0_slowdown_writes_trigger),
       OptionType::kInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, level0_slowdown_writes_trigger)}},
     {"level0_stop_writes_trigger",
-     {offsetof(struct ColumnFamilyOptions, level0_stop_writes_trigger),
+     {offset_of(&ColumnFamilyOptions::level0_stop_writes_trigger),
       OptionType::kInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, level0_stop_writes_trigger)}},
     {"max_grandparent_overlap_factor",
@@ -447,53 +461,53 @@ static std::unordered_map<std::string, OptionTypeInfo> cf_options_type_info = {
     {"max_mem_compaction_level",
      {0, OptionType::kInt, OptionVerificationType::kDeprecated, false, 0}},
     {"max_write_buffer_number",
-     {offsetof(struct ColumnFamilyOptions, max_write_buffer_number),
+     {offset_of(&ColumnFamilyOptions::max_write_buffer_number),
       OptionType::kInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_write_buffer_number)}},
     {"max_write_buffer_number_to_maintain",
-     {offsetof(struct ColumnFamilyOptions, max_write_buffer_number_to_maintain),
+     {offset_of(&ColumnFamilyOptions::max_write_buffer_number_to_maintain),
       OptionType::kInt, OptionVerificationType::kNormal, false, 0}},
     {"min_write_buffer_number_to_merge",
-     {offsetof(struct ColumnFamilyOptions, min_write_buffer_number_to_merge),
+     {offset_of(&ColumnFamilyOptions::min_write_buffer_number_to_merge),
       OptionType::kInt, OptionVerificationType::kNormal, false, 0}},
     {"num_levels",
-     {offsetof(struct ColumnFamilyOptions, num_levels), OptionType::kInt,
+     {offset_of(&ColumnFamilyOptions::num_levels), OptionType::kInt,
       OptionVerificationType::kNormal, false, 0}},
     {"source_compaction_factor",
      {0, OptionType::kInt, OptionVerificationType::kDeprecated, true, 0}},
     {"target_file_size_multiplier",
-     {offsetof(struct ColumnFamilyOptions, target_file_size_multiplier),
+     {offset_of(&ColumnFamilyOptions::target_file_size_multiplier),
       OptionType::kInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, target_file_size_multiplier)}},
     {"arena_block_size",
-     {offsetof(struct ColumnFamilyOptions, arena_block_size),
-      OptionType::kSizeT, OptionVerificationType::kNormal, true,
+     {offset_of(&ColumnFamilyOptions::arena_block_size), OptionType::kSizeT,
+      OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, arena_block_size)}},
     {"inplace_update_num_locks",
-     {offsetof(struct ColumnFamilyOptions, inplace_update_num_locks),
+     {offset_of(&ColumnFamilyOptions::inplace_update_num_locks),
       OptionType::kSizeT, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, inplace_update_num_locks)}},
     {"max_successive_merges",
-     {offsetof(struct ColumnFamilyOptions, max_successive_merges),
+     {offset_of(&ColumnFamilyOptions::max_successive_merges),
       OptionType::kSizeT, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_successive_merges)}},
     {"memtable_huge_page_size",
-     {offsetof(struct ColumnFamilyOptions, memtable_huge_page_size),
+     {offset_of(&ColumnFamilyOptions::memtable_huge_page_size),
       OptionType::kSizeT, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, memtable_huge_page_size)}},
     {"memtable_prefix_bloom_huge_page_tlb_size",
      {0, OptionType::kSizeT, OptionVerificationType::kDeprecated, true, 0}},
     {"write_buffer_size",
-     {offsetof(struct ColumnFamilyOptions, write_buffer_size),
-      OptionType::kSizeT, OptionVerificationType::kNormal, true,
+     {offset_of(&ColumnFamilyOptions::write_buffer_size), OptionType::kSizeT,
+      OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, write_buffer_size)}},
     {"bloom_locality",
-     {offsetof(struct ColumnFamilyOptions, bloom_locality),
-      OptionType::kUInt32T, OptionVerificationType::kNormal, false, 0}},
+     {offset_of(&ColumnFamilyOptions::bloom_locality), OptionType::kUInt32T,
+      OptionVerificationType::kNormal, false, 0}},
     {"memtable_prefix_bloom_bits",
      {0, OptionType::kUInt32T, OptionVerificationType::kDeprecated, true, 0}},
     {"memtable_prefix_bloom_size_ratio",
-     {offsetof(struct ColumnFamilyOptions, memtable_prefix_bloom_size_ratio),
+     {offset_of(&ColumnFamilyOptions::memtable_prefix_bloom_size_ratio),
       OptionType::kDouble, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, memtable_prefix_bloom_size_ratio)}},
     {"memtable_prefix_bloom_probes",
@@ -501,72 +515,72 @@ static std::unordered_map<std::string, OptionTypeInfo> cf_options_type_info = {
     {"min_partial_merge_operands",
      {0, OptionType::kUInt32T, OptionVerificationType::kDeprecated, true, 0}},
     {"max_bytes_for_level_base",
-     {offsetof(struct ColumnFamilyOptions, max_bytes_for_level_base),
+     {offset_of(&ColumnFamilyOptions::max_bytes_for_level_base),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_bytes_for_level_base)}},
     {"max_bytes_for_level_multiplier",
-     {offsetof(struct ColumnFamilyOptions, max_bytes_for_level_multiplier),
+     {offset_of(&ColumnFamilyOptions::max_bytes_for_level_multiplier),
       OptionType::kDouble, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_bytes_for_level_multiplier)}},
     {"max_bytes_for_level_multiplier_additional",
-     {offsetof(struct ColumnFamilyOptions,
-               max_bytes_for_level_multiplier_additional),
+     {offset_of(
+          &ColumnFamilyOptions::max_bytes_for_level_multiplier_additional),
       OptionType::kVectorInt, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions,
                max_bytes_for_level_multiplier_additional)}},
     {"max_sequential_skip_in_iterations",
-     {offsetof(struct ColumnFamilyOptions, max_sequential_skip_in_iterations),
+     {offset_of(&ColumnFamilyOptions::max_sequential_skip_in_iterations),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, max_sequential_skip_in_iterations)}},
     {"target_file_size_base",
-     {offsetof(struct ColumnFamilyOptions, target_file_size_base),
+     {offset_of(&ColumnFamilyOptions::target_file_size_base),
       OptionType::kUInt64T, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, target_file_size_base)}},
     {"rate_limit_delay_max_milliseconds",
      {0, OptionType::kUInt, OptionVerificationType::kDeprecated, false, 0}},
     {"compression",
-     {offsetof(struct ColumnFamilyOptions, compression),
+     {offset_of(&ColumnFamilyOptions::compression),
       OptionType::kCompressionType, OptionVerificationType::kNormal, true,
       offsetof(struct MutableCFOptions, compression)}},
     {"compression_per_level",
-     {offsetof(struct ColumnFamilyOptions, compression_per_level),
+     {offset_of(&ColumnFamilyOptions::compression_per_level),
       OptionType::kVectorCompressionType, OptionVerificationType::kNormal,
       false, 0}},
     {"bottommost_compression",
-     {offsetof(struct ColumnFamilyOptions, bottommost_compression),
+     {offset_of(&ColumnFamilyOptions::bottommost_compression),
       OptionType::kCompressionType, OptionVerificationType::kNormal, false, 0}},
     {"comparator",
-     {offsetof(struct ColumnFamilyOptions, comparator), OptionType::kComparator,
+     {offset_of(&ColumnFamilyOptions::comparator), OptionType::kComparator,
       OptionVerificationType::kByName, false, 0}},
     {"prefix_extractor",
-     {offsetof(struct ColumnFamilyOptions, prefix_extractor),
+     {offset_of(&ColumnFamilyOptions::prefix_extractor),
       OptionType::kSliceTransform, OptionVerificationType::kByNameAllowNull,
       false, 0}},
     {"memtable_insert_with_hint_prefix_extractor",
-     {offsetof(struct ColumnFamilyOptions,
-               memtable_insert_with_hint_prefix_extractor),
+     {offset_of(
+          &ColumnFamilyOptions::memtable_insert_with_hint_prefix_extractor),
       OptionType::kSliceTransform, OptionVerificationType::kByNameAllowNull,
       false, 0}},
     {"memtable_factory",
-     {offsetof(struct ColumnFamilyOptions, memtable_factory),
+     {offset_of(&ColumnFamilyOptions::memtable_factory),
       OptionType::kMemTableRepFactory, OptionVerificationType::kByName, false,
       0}},
     {"table_factory",
-     {offsetof(struct ColumnFamilyOptions, table_factory),
-      OptionType::kTableFactory, OptionVerificationType::kByName, false, 0}},
+     {offset_of(&ColumnFamilyOptions::table_factory), OptionType::kTableFactory,
+      OptionVerificationType::kByName, false, 0}},
     {"compaction_filter",
-     {offsetof(struct ColumnFamilyOptions, compaction_filter),
+     {offset_of(&ColumnFamilyOptions::compaction_filter),
       OptionType::kCompactionFilter, OptionVerificationType::kByName, false,
       0}},
     {"compaction_filter_factory",
-     {offsetof(struct ColumnFamilyOptions, compaction_filter_factory),
+     {offset_of(&ColumnFamilyOptions::compaction_filter_factory),
       OptionType::kCompactionFilterFactory, OptionVerificationType::kByName,
       false, 0}},
     {"merge_operator",
-     {offsetof(struct ColumnFamilyOptions, merge_operator),
+     {offset_of(&ColumnFamilyOptions::merge_operator),
       OptionType::kMergeOperator, OptionVerificationType::kByName, false, 0}},
     {"compaction_style",
-     {offsetof(struct ColumnFamilyOptions, compaction_style),
+     {offset_of(&ColumnFamilyOptions::compaction_style),
       OptionType::kCompactionStyle, OptionVerificationType::kNormal, false,
       0}}};
